requirements.txt
torch-geometric
numpy
panda
scikit-learn
jupyter
click
mlflow
yfinance
tqdm
DadosAbertosBrasil
black
streamlit
matplotlib
uniplot